style:clear main.cpp
