+++ log.txt
@@ -1,1 +1,2 @@
 01.01.1930
+02.05,2006
