+++ menu.h
@@ -1,0 +1,3 @@
+void printMenu(int, bool, char[]);
+void doSomething(int, char*, bool*);
+void menuFunc();
