+++ workWithData.h
@@ -1,0 +1,2 @@
+bool isSample(char[]);
+void Split(char[], int*);
