+++ workWithDate.h
@@ -1,0 +1,4 @@
+bool visokosYear(int);
+int valDayInMonth(int, int);
+bool isValidDate(char[]);
+
